Abstract code for better comprehension and organization
